Add go_template action, plus more
The go_template action executes the given path(s) as go templates, which
can access the user-provided template variables when doing
substitutions. The resulting template output is written back to the file
that it came from.

Also moved some functions from render.go into render_action.go per Chen's
earlier request.

For friendly error messages, there's new code that wraps the error
returned by template.Execute() to provide a friendlier error message in
the case where the user's template references a nonexistent input.
