+++ templates/commands/render_action.go
@@ -17,12 +17,18 @@ package commands
 import (
 	"bytes"
 	"errors"
+	"fmt"
+	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
 	"regexp"
+	"sort"
+	"strings"
+	"text/template"
 
 	"github.com/abcxyz/abc/templates/model"
+	"golang.org/x/exp/maps"
 )
 
 // Called with the contents of a file, and returns the new contents of the file
@@ -58,12 +64,17 @@ func walkAndModify(pos *model.ConfigPos, rfs renderFS, scratchDir, relPath strin
 			return model.ErrWithPos(pos, "Readfile(): %w", err) //nolint:wrapcheck
 		}
 
+		relToScratchDir, err := filepath.Rel(scratchDir, path)
+		if err != nil {
+			return model.ErrWithPos(pos, "Rel(): %w", err) //nolint:wrapcheck
+		}
+
 		// We must clone oldBuf to guarantee that the callee won't change the
 		// underlying bytes. We rely on an unmodified oldBuf below in the call
 		// to bytes.Equal.
 		newBuf, err := v(bytes.Clone(oldBuf))
 		if err != nil {
-			return err
+			return fmt.Errorf("when processing template file %q: %w", relToScratchDir, err)
 		}
 
 		if bytes.Equal(oldBuf, newBuf) {
@@ -99,4 +110,151 @@ func templateAndCompileRegexes(regexes []model.String, inputs map[string]string)
 	}
 
 	return compiled, merr
+}
+
+// A template parser helper to remove the boilerplate of parsing with our
+// desired options.
+func parseGoTmpl(tpl string) (*template.Template, error) {
+	return template.New("").Option("missingkey=error").Parse(tpl) //nolint:wrapcheck
+}
+
+// pos may be nil if the template is not coming from the spec file and therefore
+// there's no reason to print out spec file location in an error message. If
+// template execution fails because of a missing input variable, the error will
+// be wrapped in a unknownTemplateKeyErr.
+func parseAndExecuteGoTmpl(pos *model.ConfigPos, tmpl string, inputs map[string]string) (string, error) {
+	parsedTmpl, err := parseGoTmpl(tmpl)
+	if err != nil {
+		return "", model.ErrWithPos(pos, `error compiling as go-template: %w`, err) //nolint:wrapcheck
+	}
+
+	// As of go1.20, if the template references a nonexistent variable, then the
+	// returned error will be of type *errors.errorString; unfortunately there's
+	// no distinctive error type we can use to detect this particular error. We
+	// only get this error because we asked for Option("missingkey=error") when
+	// parsing the template. Otherwise it would silently insert "<no value>".
+	var sb strings.Builder
+	if err := parsedTmpl.Execute(&sb, inputs); err != nil {
+		// If this error looks like a missing key error, then replace it with a
+		// more helpful error.
+		matches := templateKeyErrRegex.FindStringSubmatch(err.Error())
+		if matches != nil {
+			inputKeys := maps.Keys(inputs)
+			sort.Strings(inputKeys)
+			err = &unknownTemplateKeyErr{
+				key:           matches[1],
+				availableKeys: inputKeys,
+				wrapped:       err,
+			}
+		}
+		return "", model.ErrWithPos(pos, "template.Execute() failed: %w", err) //nolint:wrapcheck
+	}
+	return sb.String(), nil
+}
+
+// unknownTemplateKeyErr is an error that will be returned when a template
+// references a variable that's nonexistent.
+type unknownTemplateKeyErr struct {
+	key           string
+	availableKeys []string
+	wrapped       error
+}
+
+func (n *unknownTemplateKeyErr) Error() string {
+	return fmt.Sprintf("the template referenced a nonexistent input variable name %q; available variable names are %v",
+		n.key, n.availableKeys)
+}
+
+func (n *unknownTemplateKeyErr) Unwrap() error {
+	return n.wrapped
+}
+
+func (n *unknownTemplateKeyErr) Is(other error) bool {
+	_, ok := other.(*unknownTemplateKeyErr)
+	return ok
+}
+
+var templateKeyErrRegex = regexp.MustCompile(`map has no entry for key "([^"]*)"`)
+
+// "srcRoot" may be a file or directory. "pos" is only used for error messages.
+func copyRecursive(pos *model.ConfigPos, srcRoot, dstRoot string, rfs renderFS, overwrite, dryRun bool) (outErr error) {
+	return fs.WalkDir(rfs, srcRoot, func(path string, de fs.DirEntry, err error) error { //nolint:wrapcheck
+		if err != nil {
+			return err // There was some filesystem error. Give up.
+		}
+
+		if de.IsDir() {
+			return nil
+		}
+
+		// We don't have to worry about symlinks here because we passed
+		// DisableSymlinks=true to go-getter.
+
+		relToSrc, err := filepath.Rel(srcRoot, path)
+		if err != nil {
+			return model.ErrWithPos(pos, "filepath.Rel(%s,%s): %w", srcRoot, path, err) //nolint:wrapcheck
+		}
+		dst := filepath.Join(dstRoot, relToSrc)
+
+		// The spec file may specify a file to copy that's deep in a
+		// directory tree, without naming its parent directory. We can't
+		// rely on WalkDir having traversed the parent directory of $path,
+		// so we must create the target directory if it doesn't exist.
+		inDir := filepath.Dir(dst)
+		if err := mkdirAllChecked(pos, rfs, inDir, dryRun); err != nil {
+			return err
+		}
+
+		dstInfo, err := rfs.Stat(dst)
+		if err == nil {
+			if dstInfo.IsDir() {
+				return model.ErrWithPos(pos, "cannot overwrite a directory with a file of the same name, %q", path) //nolint:wrapcheck
+			}
+			if !overwrite {
+				return model.ErrWithPos(pos, "destination file %s already exists and overwriting was not enabled", path) //nolint:wrapcheck
+			}
+		} else if !os.IsNotExist(err) {
+			return model.ErrWithPos(pos, "Stat(): %w", err) //nolint:wrapcheck
+		}
+
+		srcInfo, err := rfs.Stat(path)
+		if err != nil {
+			return fmt.Errorf("Stat(): %w", err)
+		}
+
+		rf, err := rfs.Open(path)
+		if err != nil {
+			return model.ErrWithPos(pos, "Open(): %w", err) //nolint:wrapcheck
+		}
+		defer func() { outErr = errors.Join(outErr, rf.Close()) }()
+
+		if dryRun {
+			return nil
+		}
+
+		// The permission bits on the output file are copied from the input file;
+		// this preserves the execute bit on executable files.
+		wf, err := rfs.OpenFile(dst, os.O_CREATE|os.O_WRONLY, srcInfo.Mode().Perm())
+		if err != nil {
+			return model.ErrWithPos(pos, "OpenFile(): %w", err) //nolint:wrapcheck
+		}
+		defer func() { outErr = errors.Join(outErr, wf.Close()) }()
+
+		if _, err := io.Copy(wf, rf); err != nil {
+			return fmt.Errorf("Copy(): %w", err)
+		}
+
+		return nil
+	})
+}
+
+// safeRelPath returns an error if the path is absolute or if it contains a ".." traversal.
+func safeRelPath(pos *model.ConfigPos, p string) error {
+	if strings.Contains(p, "..") {
+		return model.ErrWithPos(pos, `path must not contain ".."`) //nolint:wrapcheck
+	}
+	if filepath.IsAbs(p) {
+		return model.ErrWithPos(pos, "path must be relative, not absolute") //nolint:wrapcheck
+	}
+	return nil
 }
