+++ templates/commands/render_action_stringreplace_test.go
@@ -146,7 +146,7 @@ func TestActionStringReplace(t *testing.T) {
 			want: map[string]string{
 				"my_file.txt": "foo",
 			},
-			wantErr: `no entry for key "myinput"`,
+			wantErr: `nonexistent input variable name "myinput"`,
 		},
 		{
 			name:  "templated_toreplace_missing_input_should_fail",
@@ -164,7 +164,7 @@ func TestActionStringReplace(t *testing.T) {
 			want: map[string]string{
 				"my_file.txt": "foo",
 			},
-			wantErr: `no entry for key "myinput"`,
+			wantErr: `nonexistent input variable name "myinput"`,
 		},
 		{
 			name:  "templated_with_missing_input_should_fail",
@@ -182,7 +182,7 @@ func TestActionStringReplace(t *testing.T) {
 			want: map[string]string{
 				"my_file.txt": "foo",
 			},
-			wantErr: `no entry for key "myinput"`,
+			wantErr: `nonexistent input variable name "myinput"`,
 		},
 		{
 			name:  "fs_errors_should_be_returned",
